Address review feedback
- compare `container.clientWidth`/`clientHeight` instead of
  `getBoundingClientRect()` in the resize predicate: MapLibre derives the
  canvas CSS size from the integer client dimensions, so the float comparison
  made the guard a near no-op on fractionally sized containers
- cancel a pending resize animation frame before opening a new debounce
  window, so a frame queued by an earlier discrete change cannot resize the
  backing store mid-drag
- clear the window-resize burst state when a panel splitter drag starts, so
  the two resize state machines stay independent
- extract the scheduler into `packages/map/src/map-resize.ts` and use it from
  `SecondaryMapCanvas` too, so the compare/swipe panes get the same flash-free
  behaviour now that `trackResize: false` applies to every controller map
