+++ packages/map/src/map-controller.ts
@@ -580,8 +580,9 @@ export class MapController {
       renderWorldCopies: mapPreferences.renderWorldCopies,
       attributionControl: false,
       maplibreLogo: false,
-      // MapCanvas owns resizing through a ResizeObserver because its container
-      // also changes when app panels open and close. Letting MapLibre listen to
+      // The canvases own resizing through the shared scheduler in map-resize.ts
+      // (both MapCanvas and SecondaryMapCanvas) because the container also
+      // changes when app panels open and close. Letting MapLibre listen to
       // window.resize as well causes competing framebuffer reallocations while
       // a browser window is dragged, briefly exposing a transparent canvas.
       trackResize: false,
